requirement.txt
altgraph==0.17.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Flask==2.1.2
fonttools==4.33.3
future==0.18.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
packaging==21.3
pefile==2021.9.3
Pillow==9.1.0
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pyparsing==3.0.9
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
requests==2.27.1
six==1.16.0
torch==1.11.0
torchvision==0.12.0
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0